# swagger.yaml
swagger: "2.0"
info:
  description: "This is a simple API documentation for the Clicka server."
  version: 1.0.0
  title: Clicka API Documentation
  contact:
    email: "odumz20@gmail.com"
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers: 
  - "clicka.herokuapp.com"
  - "localhost:5000"
basePath: /
...
schemes:
  - http
  - https
paths:
...
  /api/v1/clients/{clientId}:
    get:
      tags:
        - client
      summary: Find client by ID
      description: Returns a single client
      operationId: getClientById
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/xml
        - application/json
      parameters:
        - name: clientId
          in: path
          description: ID of client to return
          required: true
          type: string
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/Client"
        "400":
          description: Invalid ID supplied
        "404":
          description: Client with the given ID not found
...
  /api/v1/clients/edit/{clientId}:
    put:
      tags:
        - client
      summary: Update client
      description: Update the client's details
      operationId: updateClient
      produces:
        - application/xml
        - application/json
      parameters:
        - name: clientId
          in: path
          description: ID of  client that needs to be edited
          required: true
          type: string
        - in: body
          name: body
          description: Updated details of client
          required: true
          schema:
            $ref: "#/definitions/Client"
          example:
              name: Frances
              phone: 14265789532
              provider: ['614b7036c9676b3e03e42ea8']
      responses:
        "200":
          description: successful operation
        "400":
          description: Invalid client data supplied
        "404":
          description: Client not found
...
  /api/v1/clients/delete/{clientId}:
    delete:
      tags:
        - client
      summary: Deletes a client
      description: "Delete the client by given id"
      operationId: deleteClient
      consumes:
        - multipart/form-data
        - application/x-www-form-urlencoded
      produces:
        - application/xml
        - application/json
      parameters:
        - name: clientId
          in: path
          description: Client id to delete
          required: true
          type: string
      responses:
        "200":
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Client with the given ID not found
...
definitions:
  Provider:
    required:
      - name
    properties:
      id:
        type: string
      name:
        type: string
    xml:
      name: Provider
  Client:
    required:
      - name
      - email
      - phone
      - provider
    properties:
      id:
        type: string
      name:
        type: string
      email:
        type: string
      phone:
        type: integer
      provider:
        type: array
        xml:
          name: provider
          wrapped: true
        items:
          $ref: "#/definitions/Provider"
    xml:
      name: Client
...